Add support for Mercure to VueJS
