+++ templates/react-common/utils/dataAccess.js
@@ -35,14 +35,6 @@ export function fetch(id, options = {}) {
   });
 }
 
-export function mercureSubscribe(url, topics) {
-  topics.forEach(topic =>
-    url.searchParams.append('topic', new URL(topic, ENTRYPOINT))
-  );
-
-  return new EventSource(url.toString());
-}
-
 export function normalize(data) {
   if (has(data, 'hydra:member')) {
     // Normalize items in collections
@@ -57,15 +49,4 @@ export function normalize(data) {
       ? value.map(v => get(v, '@id', v))
       : get(value, '@id', value)
   );
-}
-
-export function extractHubURL(response) {
-  const linkHeader = response.headers.get('Link');
-  if (!linkHeader) return null;
-
-  const matches = linkHeader.match(
-    /<([^>]+)>;\s+rel=(?:mercure|"[^"]*mercure[^"]*")/
-  );
-
-  return matches && matches[1] ? new URL(matches[1], ENTRYPOINT) : null;
 }
